Refactor AppIntro component and add features prop
